requirements.txt
accelerate==0.32.1
activation==0.1.0
albumentations==1.4.11
apex==0.9.10dev
appdirs==1.4.4
aria2p==0.12.0
awscli==1.33.26
bitsandbytes==0.43.1
bpy==4.1.0
cachetools==4.2.2
chumpy==0.70
clip==0.2.0
comet_ml==3.44.1
commentjson==0.9.0
ConfigArgParse==1.5.3
controlnet_aux==0.0.9
cyclonedds==0.10.5
dctorch==0.1.2
dearpygui==1.8.0
diffusers==0.29.2
docutils==0.21.2
easydict==1.13
einops==0.8.0
ema_pytorch==0.5.1
envlight==0.1.0
face_alignment==1.4.1
fastcore==1.5.54
fire==0.6.0
flash_attn==2.6.1
fpsample==0.3.2
functorch==2.0.0
fvcore==0.1.5.post20221221
gdown==5.2.0
ghapi==1.0.5
gsplat==1.0.0
h5py==3.7.0
habitat_sim==0.2.4.dev20230403
hydra-core==1.3.2
icecream==2.1.3
igl==2.2.1
imageio==2.23.0
importlib_metadata==6.0.0
InsightFace==0.7.3
iopath==0.1.10
ipdb==0.13.13
ipython==8.12.3
jaxtyping==0.2.33
joblib==1.2.0
jsonmerge==1.9.2
kornia==0.6.9
loguru==0.7.2
lpips==0.1.4
magnum==18.0.0
mathutils==3.3.0
matplotlib==3.6.3
mediapy==1.2.2
moderngl==5.10.0
multiprocess==0.70.16
natsort==8.4.0
natten==0.17.1
nbformat==5.10.4
nerfacc==0.3.3
networkx==2.8.8
newrawpy==0.20.0
nox==2024.4.15
numpy==1.23.5
omegaconf==2.2.3
onnxruntime==1.18.1
open3d==0.17.0
opencv_python==4.6.0.66
pandas==1.5.2
pathos==0.3.2
Pillow==9.3.0
Pillow==10.4.0
plotly==5.14.0
plyfile==1.0.3
projectaria_tools==1.5.2
pycolmap==0.6.1
pyembree==0.2.11
Pygments==2.11.2
Pygments==2.13.0
pyhocon==0.3.59
PyMCubes==0.1.2
pymeshlab==2023.12.post1
pyntcloud==0.3.1
pyquaternion==0.9.9
pyrender==0.1.45
pysdf==0.1.8
pytest==8.2.2
pytorch_lightning==1.9.5
pytorch_msssim==1.0.0
pytransform3d==3.5.0
PyYAML==6.0
PyYAML==6.0.1
rawpy==0.22.0
Requests==2.32.3
resize_right==0.0.2
rich==13.7.1
runstats==2.0.0
safetensors==0.4.3
scikit_image==0.19.3
scikit_learn==1.2.2
scipy==1.10.0
seaborn==0.13.2
segment_anything==1.0
setuptools==65.6.3
soda==0.0.1
sphinx_rtd_theme==2.0.0
splines==0.3.2
sshconf==0.2.7
sympy==1.13.0
tabulate==0.9.0
tensorboardX==2.5.1
tensorboardX==2.6.2.2
tensorflow==2.5.0
tensorflow_probability==0.24.0
termcolor==2.4.0
timm==1.0.7
tomesd==0.1.3
toml==0.10.2
torch_efficient_distloss==0.1.3
torch_ema==0.3
torch_scatter==2.1.2
torchdiffeq==0.2.4
torchmetrics==1.0.2
torchnet==0.0.4
torchsde==0.2.6
torchtyping==0.1.4
torchvision==0.12.0+cu113
torchviz==0.0.2
tqdm==4.64.1
transformers==4.42.4
transforms3d==0.4.2
trimesh==3.17.1
typeguard==4.3.0
typing_extensions==4.12.2
tyro==0.8.5
viz==0.0.5
xatlas==0.0.9
yacs==0.1.8
